Add a macro to generate compile time errors on struct/class sizes
Will be using this in the future when defining Skyrim records, to be sure the
structs are the same size as the sub-records, making it easy to just do a
memcpy of the data into the struct, rather than having to break the sub-
record into pieces to initalize the struct members.  This technique is already
used throughout CBash, but it relies on the structs being the correct compiled
size (packing differences in compilers will screw it up).