Removed python scripts
Finally took the time to examine what they do:
* strip.py: cleans off extra whitespace, and aligns the `= value` part of enums
* apistrip.py: checks for coding style in the various ***RecordAPI.cpp files,
  and cleans off extra whitespace

Unfortunately, both trim the trailing newline at the end of the files, probably
could be easily fixed.  In any case, not really something I feel should be a
part of the repo, as it's more of a script that someone wanting a specific
coding style would use before committing anyway.