Added _fake_ wrapper code for cint.pyd
For now, I just grabbed one of the pybind11 examples and renamed it
and removed some of the stuff that was excessive for what I needed.

The module already generates a VC++ project and compiles to PYD in the
release folder and simply publishes a few symbols that remained from
the example.