Cleaned .gitignore from pre-generated cruft
As suggested in a code comment, pre-generated .gitignores are bad.

I've replaced this file's content with what's strictly needed at this
moment for files generated from sources in this repo, or projects
files, temporary/intermediate files.

Since CMAKE generates most of the garbage in ./build directory, most
of the work is done by ignoring that folder.

(#10)