(#16) Assertion caused by an incorrect delete
`NonNullStringRecord::Read` was assigning _value to the received buffer
even when size == 0, and since the condition to call delete in the
destructor was size == 0, the code was trying to call delete[] on a
pointer that pointed to the shared memory buffer and was not valid
for the heap.

Probably this was only important on Debug mode, since on Release mode
the freeing was a NOP, and didn't failed at all.