(#18) Updated testCompress to use new fixture
Added new data files generated by the tool LZ4.exe as model for
checking  the function implementation.

As they are now, the tests don't pass because the tool writes the
output  data with a specific format generated by the 'frame'
functions, instead of the simple ones that I was publishing.